chore: separate market pallet logic to the modules
