+++ pallets/market/src/hooks/mod.rs
@@ -1,0 +1,3 @@
+mod on_finalize;
+
+pub use on_finalize::on_finalize;
