+++ pallets/market/src/hooks/on_finalize.rs
@@ -1,0 +1,89 @@
+use frame_system::pallet_prelude::*;
+use primitives::{self, configs::BalanceOf, deals::DealState};
+
+use crate::{
+    slash_and_burn, unlock_funds, Config, DealsForBlock, Event, Pallet, PendingProposals,
+    Proposals, LOG_TARGET,
+};
+
+pub fn on_finalize<T>(current_block: BlockNumberFor<T>)
+where
+    T: Config,
+{
+    let deal_ids = DealsForBlock::<T>::get(&current_block);
+    if deal_ids.is_empty() {
+        log::info!(target: LOG_TARGET, "on_finalize: no deals to process in block: {:?}", current_block);
+        return;
+    }
+
+    // INVARIANT: every deal in deal_ids is unique.
+    // PRE-COND: deal validation has been performed by `publish_storage_deals`.
+    let mut pending_proposals = PendingProposals::<T>::get();
+    for deal_id in deal_ids {
+        let Ok(proposal) = Proposals::<T>::try_get(&deal_id) else {
+            // Proposal might have been cleaned up by manual settlement or termination prior to reaching
+            // this scheduled block. Nothing more to do for this deal.
+            continue;
+        };
+
+        match &proposal.state {
+            DealState::Published => {
+                debug_assert!(
+                    proposal.start_block == current_block,
+                    "deals are scheduled to be checked only at their start block"
+                );
+
+                // Deal has not been activated, time to slash!
+                // PRE-COND: deal cannot make to this stage without being validated and proper funds allocated
+                let Some(total_storage_fee) = proposal.total_storage_fee() else {
+                    log::error!(target: LOG_TARGET, "on_finalize: invariant violated cannot calculate total storage fee, deal {}", deal_id);
+                    continue;
+                };
+                let Ok(client_fee) = TryInto::<BalanceOf<T>>::try_into(total_storage_fee) else {
+                    log::error!(target: LOG_TARGET, "on_finalize: invariant violated, cannot convert total storage to {}, deal {}", total_storage_fee, deal_id);
+                    continue;
+                };
+
+                let Ok(()) = unlock_funds::<T>(&proposal.client, client_fee) else {
+                    log::error!(target: LOG_TARGET, "on_finalize: invariant violated, failed to return the fee to the client, deal {}", deal_id);
+                    continue;
+                };
+
+                log::info!(
+                    "on_finalize: slashing {:?} for not activating a deal {}",
+                    proposal.provider,
+                    deal_id
+                );
+                // PRE-COND: deal MUST BE validated and the proper funds allocated
+                let Ok(()) = slash_and_burn::<T>(&proposal.provider, proposal.provider_collateral)
+                else {
+                    log::error!(target: LOG_TARGET, "on_finalize: invariant violated, cannot slash the deal {}", deal_id);
+                    continue;
+                };
+
+                Pallet::<T>::deposit_event(Event::<T>::DealSlashed {
+                    deal_id,
+                    provider: proposal.provider.clone(),
+                    client: proposal.client.clone(),
+                    amount: proposal.provider_collateral,
+                });
+            }
+            DealState::Active(_) => {
+                log::info!(
+                    "on_finalize: deal {} has been properly activated before, all good.",
+                    deal_id
+                );
+                continue;
+            }
+        }
+
+        // Deal has been processed, no need to process it twice.
+        Proposals::<T>::remove(&deal_id);
+        // PRE-COND: all deals in DealsPerBlock are published.
+        // All Published deals are hashed and added to [`PendingProposals`].
+        let _ = pending_proposals.remove(&Pallet::<T>::hash_proposal(&proposal));
+    }
+
+    PendingProposals::<T>::set(pending_proposals);
+    DealsForBlock::<T>::remove(&current_block);
+}
